requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
pywebview>=6.0
